refactor: separate init-config command.
